requirements.txt
streamlit
openai-whisper
moviepy
setuptools